+++ src/function.rs
@@ -27,6 +27,9 @@ pub struct FunctionProperties {
     /// List of all named arguments for this function.
     pub named_arguments: Vec<WString>,
 
+    /// Whether to give an error if the function is called with an incorrect number of arguments
+    pub strict_arity: bool,
+
     /// The index of the named_arguments that ends in ...
     pub variadic: Option<usize>,
 
@@ -477,8 +480,12 @@ impl FunctionProperties {
         }
 
         let named = &self.named_arguments;
-        if !named.is_empty() {
-            sprintf!(=> &mut out, " --argument-names");
+        if !named.is_empty() || self.strict_arity {
+            sprintf!(=> &mut out, if self.strict_arity {
+                " --strict-argument-names"
+            } else {
+                " --argument-names"
+            });
             for (idx, name) in named.iter().enumerate() {
                 sprintf!(=> &mut out, " %ls", name);
                 if self.variadic == Some(idx) {